+++ src/Command/PortalNode/Extension/DeactivateExtension.php
@@ -6,7 +6,7 @@ namespace Heptacom\HeptaConnect\Bridge\ShopwarePlatform\Command\PortalNode\Exten
 
 use Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Deactivate\PortalExtensionDeactivatePayload;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalExtension\PortalExtensionDeactivateActionInterface;
-use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
+use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeySerializerContract;
 use Heptacom\HeptaConnect\Storage\Base\Exception\UnsupportedStorageKeyException;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
@@ -20,7 +20,7 @@ class DeactivateExtension extends Command
     use ExtensionCommandTrait;
 
     public function __construct(
-        private StorageKeyGeneratorContract $storageKeyGenerator,
+        private StorageKeySerializerContract $storageKeySerializer,
         private PortalExtensionDeactivateActionInterface $extensionDeactivateAction
     ) {
         parent::__construct();
@@ -49,7 +49,7 @@ class DeactivateExtension extends Command
             $io->success(\sprintf(
                 'Extension "%s" is now deactivated for portal-node "%s"',
                 $extensionClass,
-                $this->storageKeyGenerator->serialize($portalNodeKey)
+                $this->storageKeySerializer->serialize($portalNodeKey)
             ));
 
             return 0;
@@ -57,7 +57,7 @@ class DeactivateExtension extends Command
         $io->error(\sprintf(
             'Could not deactivate extension "%s" for portal-node "%s"',
             $extensionClass,
-            $this->storageKeyGenerator->serialize($portalNodeKey)
+            $this->storageKeySerializer->serialize($portalNodeKey)
         ));
 
         return 2;